+++ vstgui/uidescription/platformsupport.h
@@ -69,7 +69,7 @@ public:
 		kResizable  = 1 << 1,
 	};
 	
-	static PlatformWindow* create (const CRect& size, const char* title = 0, WindowType type = kPanelType, long styleFlags = 0, IPlatformWindowDelegate* delegate = 0, void* parentWindow = 0);
+	static PlatformWindow* create (const CRect& size, const char* title = 0, WindowType type = kPanelType, int32_t styleFlags = 0, IPlatformWindowDelegate* delegate = 0, void* parentWindow = 0);
 	
 	virtual void* getPlatformHandle () const = 0;
 	virtual void show () = 0;
